+++ src/main/java/nextstep/subway/line/application/dto/request/SectionRequest.java
@@ -1,10 +1,12 @@
-package nextstep.subway.line.application.dto;
+package nextstep.subway.line.application.dto.request;
 
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import javax.validation.constraints.Min;
 
 @Getter
+@NoArgsConstructor
 public class SectionRequest {
     @Min(1)
     private long upStationId;
@@ -12,9 +14,6 @@ public class SectionRequest {
     private long downStationId;
     @Min(0)
     private int distance;
-
-    public SectionRequest() {
-    }
 
     public SectionRequest(Long upStationId, Long downStationId, int distance) {
         this.upStationId = upStationId;
