--- src/main/java/nextstep/subway/station/applicaion/dto/StationResponse.java
@@ -1,22 +1,0 @@
-package nextstep.subway.station.applicaion.dto;
-
-public class StationResponse {
-    private Long id;
-    private String name;
-
-    public StationResponse() {
-    }
-
-    public StationResponse(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
