+++ src/main/java/nextstep/subway/station/applicaion/dto/request/StationRequest.java
@@ -1,10 +1,12 @@
-package nextstep.subway.station.applicaion.dto;
+package nextstep.subway.station.applicaion.dto.request;
 
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import javax.validation.constraints.NotBlank;
 
 @Getter
+@NoArgsConstructor
 public class StationRequest {
     @NotBlank
     private String name;
