+++ src/main/java/nextstep/subway/station/ui/StationController.java
@@ -2,8 +2,8 @@ package nextstep.subway.station.ui;
 
 import lombok.RequiredArgsConstructor;
 import nextstep.subway.station.applicaion.StationService;
-import nextstep.subway.station.applicaion.dto.StationRequest;
-import nextstep.subway.station.applicaion.dto.StationResponse;
+import nextstep.subway.station.applicaion.dto.request.StationRequest;
+import nextstep.subway.station.applicaion.dto.response.StationResponse;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
